+++ src/main/java/com/teacher/management/repository/CoursesManualsRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface CoursesManualsRepository extends JpaRepository<CoursesManuals,Long> {
-    public CoursesManuals findByIdCourseManual(int idCourseManual);
+    public CoursesManuals findByIdCourseManual(long idCourseManual);
     public CoursesManuals save(CoursesManuals coursesManuals);
     public List<CoursesManuals> findAll();
     public List<CoursesManuals> getCoursesManualsByTipCursManual(String tipCursManual);
