+++ src/main/java/com/teacher/management/repository/TeachingActivityRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface TeachingActivityRepository extends JpaRepository<TeachingActivity,Long> {
-    public TeachingActivity findByIdActivity(int idTeachingActivity);
+    public TeachingActivity findByIdActivity(long idTeachingActivity);
     public TeachingActivity save(TeachingActivity teachingActivity);
     public List<TeachingActivity> findAll();
     public List<TeachingActivity> getTeachingActivitiesByTipActivitate(String tipActivitate);
