+++ src/main/java/com/teacher/management/service/CoursesManualsService.java
@@ -5,7 +5,7 @@ import com.teacher.management.models.CoursesManuals;
 import java.util.List;
 
 public interface CoursesManualsService {
-    public CoursesManuals findByIdCourseManual(int idCourseManual);
+    public CoursesManuals findByIdCourseManual(long idCourseManual);
     public CoursesManuals save(CoursesManuals coursesManuals);
     public List<CoursesManuals> findAll();
     public List<CoursesManuals> getCoursesManualsByTipCursManual(String tipCursManual);
