+++ src/main/java/com/teacher/management/service/ThesisCoordinationService.java
@@ -5,7 +5,7 @@ import com.teacher.management.models.ThesisCoordination;
 import java.util.List;
 
 public interface ThesisCoordinationService {
-    public ThesisCoordination findByIdThesisCoord(int idThesisCoord);
+    public ThesisCoordination findByIdThesisCoord(long idThesisCoord);
     public ThesisCoordination save(ThesisCoordination thesisCoordination);
     public List<ThesisCoordination> findAll();
     public List<ThesisCoordination> getThesisCoordonationByTipTeza(String tipTeza);
