+++ src/main/java/com/teacher/management/service/WorkActivityService.java
@@ -5,7 +5,7 @@ import com.teacher.management.models.WorkActivity;
 import java.util.List;
 
 public interface WorkActivityService {
-    public WorkActivity findByIdWorkActivity(int idWorkActivity);
+    public WorkActivity findByIdWorkActivity(long idWorkActivity);
     public WorkActivity save(WorkActivity workActivity);
     public List<WorkActivity> findAll();
